+++ Changelog.cs
@@ -79,6 +79,7 @@ public class Changelog : Window {
             .Sub("You can also set a permanent decoration to be used all year round.");
         Add("1.10.0.0", "Support for Dawntrail");
         Add("1.10.0.1", "Updated more tweaks for Dawntrail");
+        Add("1.10.5.0", "Fixed tweak preview images not loading correctly");
     }
 
 #if DEBUG