+++ SimpleTweaksPluginConfig.cs
@@ -57,7 +57,6 @@ public partial class SimpleTweaksPluginConfig : IPluginConfiguration {
     public bool ShowOtherTweaksTab = true;
     public bool NoCallerInLog;
     public bool UseFuzzyTweakSearch = true;
-    public bool DismissedUpdateNotice;
 
     public bool ShowTweakDescriptions = true;
     public bool ShowTweakIDs;