+++ Tweaks/UiAdjustment/TimerOnDutyWaiting.cs
@@ -1,18 +1,21 @@
-﻿using System;
+using System;
 using System.Diagnostics;
+using System.Text;
 using FFXIVClientStructs.FFXIV.Client.System.String;
 using FFXIVClientStructs.FFXIV.Component.GUI;
 using Lumina.Excel.Sheets;
 using SimpleTweaksPlugin.Events;
 using SimpleTweaksPlugin.TweakSystem;
+using SeStringBuilder = Lumina.Text.SeStringBuilder;
 
 namespace SimpleTweaksPlugin.Tweaks.UiAdjustment;
 
 [TweakName("Timer on Duty Waiting")]
 [TweakDescription("Shows the 45 second countdown after readying for a duty.")]
+[Changelog("1.10.5.0", "Fixed display of text in french clients.")]
+[Changelog("1.10.5.1", "Fixed display on all clients...")]
 public unsafe class TimerOnDutyWaiting : UiAdjustments.SubTweak {
     private readonly Utf8String* timeRemainingString = Utf8String.CreateEmpty();
-    private readonly string prefix = Service.Data.GetExcelSheet<Addon>().GetRow(2780).Text.ExtractText();
     private Stopwatch stopwatch = new();
     private TimeSpan timeOffset = TimeSpan.Zero;
 
@@ -33,8 +36,10 @@ public unsafe class TimerOnDutyWaiting : UiAdjustments.SubTweak {
     private void OnPostUpdate(AtkUnitBase* addon) {
         var timeRemaining = timeOffset.Seconds - stopwatch.Elapsed.Seconds;
         if (timeRemaining <= 0) return;
-
-        timeRemainingString->SetString($"{prefix} ({timeRemaining})");
+        var seStringBuilder = new SeStringBuilder();
+        seStringBuilder.Append(Service.Data.GetExcelSheet<Addon>().GetRow(2780).Text);
+        seStringBuilder.Append($" ({timeRemaining})");
+        timeRemainingString->SetString(seStringBuilder.GetViewAsSpan());
         addon->GetTextNodeById(3)->SetText(timeRemainingString->StringPtr);
     }
 }